feat(hitbox/phase2.3): enemy collision LOD stage integration + pixel-perfect stage baseline
- Added enemy profile LOD adaptation stage to collision pipeline (maps higher importance → negative bias for prioritization).
- Implemented baseline pixel-perfect stage pruning logic:
	• FAST tier: prune every other candidate (cheap heuristic)
	• BALANCED tier: single center sample degeneracy filter
	• PRECISE/ULTRA tiers: 3x3 sampling for early discard of empty/degenerate masks
- Updated bias heuristic (importance → negative lod_bias) and stabilized ordering expectations.
- New integration test `test_collision_pipeline_phase2_3_enemy_lod` validating bias ordering across contrasting candidates (size + distance).
- Roadmap (implementation_hitbox_rework.txt): Mark Stage 3 (pixel-perfect baseline) complete with descriptive annotation.
- README: Milestone 2.3 section now reflects LOD stage + pixel-perfect baseline implemented; deferred list trimmed to advanced overlap, multi-res selection, SIMD + full mask intersection.
- No runtime gameplay paths invoke pipeline yet (feature still sandboxed); low regression risk.

+++ src/game/collision_pipeline.c
@@ -552,7 +552,8 @@ bool rogue_collision_stage_enemy_profile_lod(struct RogueCollisionContext* ctx,
         /* Distance proxy: actual distance to view center (player surrogate). */
         float dx = c->x - cx;
         float dy = c->y - cy;
-        float dist = (float) sqrtf(dx * dx + dy * dy);
+        /* MSVC C mode lacks sqrtf; use double sqrt then cast (portable). */
+        float dist = (float) sqrt((double) (dx * dx + dy * dy));
         /* Screen area proxy: treat 2*half_w * 2*half_h as projected box (no scaling). */
         float screen_area = (c->half_w * 2.f) * (c->half_h * 2.f);
         rogue_enemy_collision_profile_adapt_lod(c->enemy_profile, dist, screen_area, 0.f, 0.f);
